third_party/blink/web_tests/fast/dom/webtiming-navigate-within-document-expected.txt

This test checks that navigating within the document does not reset Web Timing numbers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS connectEnd is unchanged.
PASS connectStart is unchanged.
PASS domComplete is unchanged.
PASS domContentLoadedEventEnd is unchanged.
PASS domContentLoadedEventStart is unchanged.
PASS domInteractive is unchanged.
PASS domLoading is unchanged.
PASS domainLookupEnd is unchanged.
PASS domainLookupStart is unchanged.
PASS fetchStart is unchanged.
PASS loadEventEnd is unchanged.
PASS loadEventStart is unchanged.
PASS navigationStart is unchanged.
PASS redirectEnd is unchanged.
PASS redirectStart is unchanged.
PASS requestStart is unchanged.
PASS responseEnd is unchanged.
PASS responseStart is unchanged.
PASS secureConnectionStart is unchanged.
PASS toJSON is unchanged.
PASS unloadEventEnd is unchanged.
PASS unloadEventStart is unchanged.
PASS successfullyParsed is true

TEST COMPLETE
